Address review feedback (micheal-o):
1. Narrow exception catch: catch SparkException only when its cause
   message contains "Cannot find endpoint" (the specific
   RpcEndpointNotFoundException indicator), instead of catching all
   SparkException.
2. Remove standalone StateStoreCoordinatorBatchReadSuite.
3. Move unit test to existing StateStoreCoordinatorSuite.
4. Add integration test to StateStoreCoordinatorStreamingSuite
   (StateDataSourceReadSuite cannot access private[state] stop()).
